Add cost to table

+++ back-end/data/migrations/01_users.js
@@ -7,7 +7,7 @@ exports.up = function(knex, Promise) {
 		table.text('lastName').notNullable();
 		table.boolean('isAdmin').notNullable().defaultTo(false);
 		table.string('phone').notNullable();
-		table.integer('cost').notNullable();
+		table.decimal('cost').notNullable();
 		table.boolean('emailSubscribe').notNullable().defaultTo(false);
 		table.boolean('textSubscribe').notNullable().defaultTo(false);
 		table.text('application');
